correction on rate screen path & structure rework
