+++ src/screens/Onboarding/Kitsu/index.js
@@ -1,5 +1,4 @@
-import RateScreen from './RateScreen';
 import ImportLibrary from './ImportLibrary';
 import ImportDetail from './ImportDetail';
 
-export { RateScreen, ImportLibrary, ImportDetail };
+export { ImportLibrary, ImportDetail };
