+++ src/screens/Onboarding/common/RateScreen/index.js
@@ -1,0 +1,5 @@
+import RateScreen from './screen';
+
+export {
+  RateScreen,
+};
