+++ src/screens/Onboarding/common/index.js
@@ -2,5 +2,6 @@ import { OnboardingHeader } from './OnboardingHeader';
 import FavoritesScreen from './FavoritesScreen';
 import ManageLibrary from './ManageLibrary';
 import RatingSystemScreen from './RatingSystemScreen';
+import { RateScreen } from './RateScreen';
 
-export { OnboardingHeader, FavoritesScreen, ManageLibrary, RatingSystemScreen };
+export { OnboardingHeader, FavoritesScreen, ManageLibrary, RatingSystemScreen, RateScreen };
