+++ src/screens/Onboarding/index.js
@@ -1,6 +1,6 @@
 import { CreateAccountScreen, SelectAccountScreen, WelcomeScreen } from './Aozora';
-import { RateScreen, ImportLibrary, ImportDetail } from './Kitsu';
-import { OnboardingHeader, FavoritesScreen, ManageLibrary, RatingSystemScreen } from './common';
+import { ImportLibrary, ImportDetail } from './Kitsu';
+import { RateScreen, OnboardingHeader, FavoritesScreen, ManageLibrary, RatingSystemScreen } from './common';
 
 export {
   CreateAccountScreen,
